requirements.txt
click==8.1.2
colorama==0.4.4
Flask==2.1.1
greenlet==1.1.2
IMDbPY==2021.4.18
itsdangerous==2.1.2
Jinja2==3.1.1
lxml==4.8.0
MarkupSafe==2.1.1
pymongo==4.1.1
SQLAlchemy==1.4.35
Werkzeug==2.1.1
